+++ packages/coding-agent/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.4.2] - 2026-05-29
 
 ### Added